extract types from routes #20
