+++ packages/frontend/src/util/routes.tsx
@@ -1,38 +1,7 @@
 import * as React from 'react';
 import { HomePage } from '../components/pages/HomePage';
 import { LinkPage } from '../components/pages/LinkPage';
-import { RouteProps } from 'react-router';
-
-// If necessary, add support for: H.LocationDescriptor | ((location: H.Location) => H.LocationDescriptor);
-type LinkType = string;
-
-/**
- * Each Approute can have a specific link (i.e., path with filled parameter placeholders),
- * a display Name, i.e., text of the link and a nonoptional (!) path
- *
- * TODO: move to types package to be able to move app routes to their own
- * individual packages
- */
-export interface AppRouteProps extends RouteProps {
-  // Use this if the link target differs from the path specification,
-  // i.e., if the path url contains paramter specifications etc
-  readonly link?: LinkType;
-
-  // link text (Human readable!)
-  readonly displayName?: string;
-
-  // AppRoutes must have a path - deoptionalize this property
-  readonly path: AppPath;
-}
-
-/**
- * All available paths in this app
- */
-export enum AppPath {
-  HOME = '/home',
-  LINK = '/link',
-  ERROR = '/',
-}
+import { AppPath, AppRoute, LinkType } from 'elite-types';
 
 /**
  * Route for the Home page of this app
@@ -66,7 +35,7 @@ export const APP_ROUTES: AppRouteProps[] = [HOME_ROUTE, LINK_ROUTE];
  *
  * @param route
  */
-export function getLinkForRoute(route: AppRouteProps): LinkType {
+export function getLinkForRoute(route: AppRoute): LinkType {
   return route.link || route.path;
 }
 
@@ -89,7 +58,7 @@ export function getLinkForPath(path: AppPath): LinkType {
  *
  * @param route
  */
-export function getDisplayNameForRoute(route: AppRouteProps): string {
+export function getDisplayNameForRoute(route: AppRoute): string {
   return route.displayName || getLinkForRoute(route);
 }
