+++ packages/types/src/routes/appPath.type.ts
@@ -1,0 +1,8 @@
+/**
+ * All available paths in this app
+ */
+export enum AppPath {
+  HOME = '/home',
+  LINK = '/link',
+  ERROR = '/',
+}
