+++ packages/types/src/routes/index.ts
@@ -1,0 +1,3 @@
+export * from './appPath.type';
+export * from './appRoute.type';
+export * from './link.type';
