+++ packages/types/src/routes/link.type.ts
@@ -1,0 +1,1 @@
+export type LinkType = string;
